transceiver/requirements.txt
aiohttp==3.8.4
anyio==3.6.2
fastapi==0.95.0
flake8==7.0.0
gunicorn==20.1.0
httpx==0.23.3
loguru==0.6.0
pydantic==1.10.7
python-dotenv==1.0.0
requests==2.21.0
uvicorn[standart]==0.21.1
